backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
pydantic-settings==2.5.2
python-multipart==0.0.12
websockets==13.1
httpx==0.27.2
aiofiles==24.1.0

sqlalchemy==2.0.35
alembic==1.13.3
psycopg[binary]==3.2.3
asyncpg==0.29.0
greenlet==3.1.1

faster-whisper==1.0.3
numpy==1.26.4
soundfile==0.12.1
av>=11.0.0,<13.0.0

opencv-python-headless==4.10.0.84
mediapipe==0.10.18
deepface==0.0.93
tf-keras==2.17.0
Pillow==10.4.0

scikit-learn==1.5.2
xgboost==2.1.1
pandas==2.2.3
joblib==1.4.2

minio==7.2.9

structlog==24.4.0
python-jose[cryptography]==3.3.0

aiortc==1.9.0
